\name{genea140}
\docType{data}
\alias{genea140}
\title{Genealogical information for 140 individuals from the Quebec Reference Sample}
\description{
A genealogical corpus made of 41523 individuals from the province of Quebec, Canada. A total of 140 individuals have been sampled in seven sub-populations, listed in pop140, and their genealogies were reconstructed as far back as possible using the BALSAC population register and the Early Quebec Population Register.
}
\usage{data(genea140)}
\format{
A data frame with 41523 observations on 11 variables.

[,1] ind: An individual's ID number

[,2] father: This indidividual's father ID number

[,3] mother: This indidividual's mother ID number

[,4] sex: Individual's sex coded 1/2 for male/female
}
\source{Balsac}
\references{
http://www.quebecgenpop.ca

Roy-Gagnon, M.-H., Moreau, C., Bherer, C., St-Onge, P., Sinnett, D., Laprise, C., Vezina, H., Labuda, D. (2011). Genomic and genealogical investigation of the French Canadian founder population structure. Human Genetics, 129(5), 521-31. 
}
\keyword{datasets}
