\name{geneaJi}
\docType{data}
\alias{geneaJi}
\title{Highly inbred pedigree}
\description{
A modified version of a pedigree of two Jicaque Indians studied by Chapman & Jacquard (1971).
}
\usage{data(geneaJi)}
\format{
A data frame with 29 observations on 4 variables.

[, 1] 	ind 	An individual's ID number

[, 2] 	father 	This indidividual's father ID number

[, 3] 	mother 	This indidividual's mother ID number

[, 4] 	sex 	Individual's sex coded 1/2 for male/female
}
\source{CHAPMAN & JACQUART (1971)}
\references{
Chapman, A.M., & Jacquard, A. (1971). Un isolat d'Amerique centrale: Les indiens Jicaques du Honduras. Paris: Presses universitaires de France.
}
\keyword{datasets}
