\name{GenlibR-package}
\alias{GenlibR-package}
\docType{package}
\title{
Genealogical Data Analysis
}
\description{
Genealogical data analysis including descriptive statistics (e.g., kinship and inbreeding coefficients) and gene-dropping simulations.
}
\details{
\tabular{ll}{
Package: \tab GenlibR\cr
Type: \tab Package\cr
Version: \tab 1.1.4\cr
Date: \tab 2021-09-15\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Louis Houde, Jean-Francois Lefebvre, Valery Roy-Lagace, Sebastien Lemieux 
}
\keyword{ package }

