\name{GLmultiList-class}
\Rdversion{1.1}
\docType{class}
\alias{GLmultiList-class}

\title{Class \code{"GLmultiList"}}
\description{
Objects created to carry information mostly between invisible functions.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("GLmultiList", 'Array')}.
}
\section{Slots}{
  \describe{
    \item{\code{.Data}:}{Object of class \code{"list"} ~~ }
    \item{\code{liste}:}{Object of class \code{"list"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{list}"}, from data part.
Class \code{"\linkS4class{vector}"}, by class "list", distance 2.
}
\section{Methods}{
No methods defined with class "GLmultiList" in the signature.
}

\examples{
showClass("GLmultiList")
}
\keyword{classes}
