\name{pop140}
\docType{data}
\alias{pop140}
\title{Population of origin of the 140 Quebec samples}
\description{
  The 140 individuals from the genealogical corpus from Quebec were sampled from 7 different populations from 5 regions; Quebec City, Montreal, Saguenay, North Shore, Gaspesia. In Gaspesia we find 3 different populations: French-Canadians, Acadians and Loyalists.
  }
\usage{data(pop140)}
\format{
  [, 1]  ind:  An individual's ID number
  
  [, 2]  pop:  This indidividual's population
}
\source{Balsac}
\references{
  http://www.quebecgenpop.ca/home.html
  
  Roy-Gagnon, M.-H., Moreau, C., Bherer, C., St-Onge, P., Sinnett, D., Laprise, C., Vezina, H., Labuda, D. (2011). Genomic and genealogical investigation of the French Canadian founder population structure. Human Genetics, 129(5), 521-31.
}
\keyword{datasets}
