\name{GenlibR-package}
\alias{GenlibR-package}
\docType{package}
\title{
Genealogical Data Analysis
}
\description{
Takes genealogical data frames and calculates different demographic variables such as genetic contribution, kinship, etc...
}
\details{
\tabular{ll}{
Package: \tab GenlibR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-04-04\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Louis Houde, Jean-Francois Lefebvre, Valery Roy-Lagace, Sebastien Lemieux 
}
\keyword{ package }

